vuejs/pyproject.toml
[tool.poetry]
name = "vuejs"
version = "0.1.0"
description = ""
authors = ["Adam Hill <github@adamghill.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.3"
djangorestframework = "^3.12.4"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.5b1", allow-prereleases = true}
poethepoet = "^0.10.0"

[tool.poe.tasks]
m = "./manage.py"
r = "./manage.py runserver 0:8000"
ma = "./manage.py makemigrations"
mi = "./manage.py migrate"
md = ["ma", "mi"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
